Puzzle game
Made for fun on the coding and also on playing. I saw a puzzle like that today and remembered that as a kid i could never solve it, but now it's really easy. Filled with nostalgia i coded this and played it a lot. It was great to train java.awt and java.swing, two libs which i not very experienced yet.